GROUNDWORK-1688 Organize transit types
